scons: updated to version 0.95.5 and added selectVersion dispatcher
Scons is not only a tool, it's also a python library. However, the
modules are installed in ~/.nix-profile/lib/scons-${version} -- not in
python's "site-packages" directory -- so python won't find the library
without further help. Maybe there should be a symlink in nix-profile? I
guess there should, but I didn't know how to write that into the
expression.


git-svn-id: https://svn.nixos.org/repos/nix/nixpkgs/trunk@12157 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb