emacs: Link successfully on 64 bit.
I'm using this expression on a Linux 64-bit installation that's been patched
to use glibc 2.5, not the current version 2.7. My installation needs this
patch and it looks sensible enough to me, but I've been told that others can
install Emacs on 64 bit machines without this. Hopefully, this doesn't break
anything.


git-svn-id: https://svn.nixos.org/repos/nix/nixpkgs/trunk@12303 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb